remove unnecessary testing on VerifyUsingFallbackSerializer
